Move co2signal models to their own file
